Major DB changes

+++ listings/models.py
@@ -2,6 +2,7 @@ from django.db import models
 from django.utils.timezone import now
 from realtors.models import Realtor
 
+
 class Listing(models.Model):
 
     class SaleType(models.TextChoices):
@@ -26,7 +27,7 @@ class Listing(models.Model):
     sale_type = models.CharField(max_length=50, choices=SaleType.choices, default=SaleType.FOR_SALE)
     price = models.IntegerField()
     bedrooms = models.IntegerField()
-    bathrooms = models.DecimalField(max_digits=2, decimal_places=1)
+    bathrooms = models.IntegerField()
 
     home_type = models.CharField(max_length=50, choices=HomeType.choices, default=HomeType.HOUSE)
     sqft = models.IntegerField()
